Get (autogenerated) subject id from database, use it for reminder insert.
Pretty short message but in practice involve quite a bit of code change.
This works but could use some code clean up later.

+++ app/src/main/java/black/old/spacedrepetitionowl/SubjectRecyclerViewAdapter.kt
@@ -152,7 +152,7 @@ class SubjectRecyclerViewAdapter(
         notifyDataSetChanged()
     }
 
-    private fun getRemindersListBySubject(subjectId: Int) : List<Reminder> {
+    private fun getRemindersListBySubject(subjectId: Long) : List<Reminder> {
         // Only get Reminders from the list that has the current subject ID
         var currentReminders = reminders.filter {
             it.subjectId == subjectId
